fix: rename file to sca_policy.hpp
